Create Spider class and remove Pandoc integration
All scrapers now use a common Spider class. This class is an Abstract
Base Class, which requires subclasses to implement a parse() function
and define a domain property. The parse() function of a yields lxml
Elements. At any time during parsing a spider may update self.metadata
which is added to the final HTML document's <head> as <title> and <meta>
tags.

The domain property is used to automatically select the correct spider
in the new common fiction-scraper program. A URL may be passed to
fiction-scraper and the story will be downloaded and the scraped HTML
will be written to standard output.

All spiders are located under scraper.sites now. All spiders are loaded
via the __init__.py of scraper using some pkgutil magic.

Regression: it is no longer possible to set Pandoc command line
arguments. Notably --epub-chapter-level cannot be set anymore.