Spider: add filter list
Spiders now have a list of filter functions that are automatically
applied to the <body> element after parsing. Each filter is a simple
function that takes an lxml html element as an input and modifies it in
place. The lxml html cleaner is set as a default filter for all spiders.